Use php-coveralls phar file from v0.7.0 release.
Removed the top level composer.json and started to use the phar file
from php-coveralls 0.7.0 release.

A phar file contains all the dependencies, so no need to have
composer.json anymore. Tools like php-coveralls often depend on legacy
libs, which may conflict with dependencies of modern applications, so it
is better to use the phar file.